Create DissolutionDirectorService, Getter and Patcher Refactor
