+++ src/main/java/uk/gov/companieshouse/service/dissolution/DissolutionService.java
@@ -3,13 +3,10 @@ package uk.gov.companieshouse.service.dissolution;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.exception.DissolutionNotFoundException;
-import uk.gov.companieshouse.model.db.dissolution.Dissolution;
 import uk.gov.companieshouse.model.dto.companyofficers.CompanyOfficer;
 import uk.gov.companieshouse.model.dto.companyprofile.CompanyProfile;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionCreateRequest;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionCreateResponse;
-import uk.gov.companieshouse.model.dto.dissolution.DissolutionDirectorPatchRequest;
-import uk.gov.companieshouse.model.dto.dissolution.DissolutionDirectorPatchResponse;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionGetResponse;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionPatchRequest;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionPatchResponse;
@@ -45,10 +42,6 @@ public class DissolutionService {
         return patcher.addDirectorApproval(companyNumber, userId, body);
     }
 
-    public DissolutionDirectorPatchResponse updateSignatory(String companyNumber, DissolutionDirectorPatchRequest body, String directorId) throws DissolutionNotFoundException{
-        return patcher.updateSignatory(companyNumber, body, directorId);
-    }
-
     public void handlePayment(PaymentPatchRequest body, String applicationReference) throws DissolutionNotFoundException {
         patcher.handlePayment(body, applicationReference);
     }
@@ -71,20 +64,5 @@ public class DissolutionService {
 
     public boolean isDirectorPendingApproval(String companyNumber, String officerId) {
         return getter.isDirectorPendingApproval(companyNumber, officerId);
-    }
-
-    public boolean doesDirectorExist(String companyNumber, String officerId) {
-        return getter.doesDirectorExist(companyNumber, officerId);
-    }
-
-    public Optional<String> checkPatchDirectorConstraints(String companyNumber, String directorId, String email) throws DissolutionNotFoundException {
-        final Dissolution dissolution = repository.findByCompanyNumber(companyNumber).orElseThrow(DissolutionNotFoundException::new);
-        if (!getter.isDirectorPendingApprovalForDissolution(directorId, dissolution)) {
-            return Optional.of(DIRECTOR_IS_NOT_PENDING_APPROVAL);
-        }
-        if (!getter.doesEmailBelongToApplicant(email, dissolution)) {
-            return Optional.of(ONLY_THE_APPLICANT_CAN_UPDATE_SIGNATORY);
-        }
-        return Optional.empty();
     }
 }
