+++ src/test/java/uk/gov/companieshouse/service/dissolution/DissolutionServiceTest.java
@@ -8,13 +8,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.exception.DissolutionNotFoundException;
 import uk.gov.companieshouse.fixtures.CompanyProfileFixtures;
 import uk.gov.companieshouse.fixtures.DissolutionFixtures;
-import uk.gov.companieshouse.model.db.dissolution.Dissolution;
 import uk.gov.companieshouse.model.dto.companyofficers.CompanyOfficer;
 import uk.gov.companieshouse.model.dto.companyprofile.CompanyProfile;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionCreateRequest;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionCreateResponse;
-import uk.gov.companieshouse.model.dto.dissolution.DissolutionDirectorPatchRequest;
-import uk.gov.companieshouse.model.dto.dissolution.DissolutionDirectorPatchResponse;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionGetResponse;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionPatchRequest;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionPatchResponse;
@@ -28,8 +25,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.companieshouse.fixtures.CompanyOfficerFixtures.generateCompanyOfficer;
-import static uk.gov.companieshouse.fixtures.DissolutionDirectorFixtures.generateDissolutionPatchDirectorRequest;
-import static uk.gov.companieshouse.fixtures.DissolutionFixtures.generateDissolution;
 import static uk.gov.companieshouse.fixtures.DissolutionFixtures.generateDissolutionPatchRequest;
 import static uk.gov.companieshouse.fixtures.PaymentFixtures.generatePaymentPatchRequest;
 
@@ -158,24 +153,6 @@ public class DissolutionServiceTest {
     }
 
     @Test
-    void updateSignatory_updatesSignatory_returnsPatchResponse() throws DissolutionNotFoundException {
-        final DissolutionDirectorPatchRequest body = generateDissolutionPatchDirectorRequest();
-        body.setEmail(EMAIL);
-        body.setOnBehalfName(ON_BEHALF_NAME);
-
-        final DissolutionDirectorPatchResponse response = DissolutionFixtures.generateDissolutionDirectorPatchResponse();
-
-        when(patcher.updateSignatory(COMPANY_NUMBER, body, EMAIL)).thenReturn(response);
-
-        final DissolutionDirectorPatchResponse result = service.updateSignatory(COMPANY_NUMBER, body, EMAIL);
-
-        verify(patcher).updateSignatory(COMPANY_NUMBER, body, EMAIL);
-
-        assertNotNull(result);
-        assertEquals(response, result);
-    }
-
-    @Test
     void hasDirectorAlreadyApproved_callsDissolutionGetter_isDirectorPendingApproval() {
         final String companyNumber = "12345678";
         final String email = "user@mail.com";
@@ -190,74 +167,11 @@ public class DissolutionServiceTest {
     }
 
     @Test
-    void doesDirectorExist_callsDissolutionGetter_doesDirectorExist() {
-        final String companyNumber = "12345678";
-        final String officerId = OFFICER_ID;
-
-        when(getter.doesDirectorExist(companyNumber, officerId)).thenReturn(true);
-
-        final boolean result = service.doesDirectorExist(companyNumber, officerId);
-
-        verify(getter).doesDirectorExist(companyNumber, officerId);
-
-        assertTrue(result);
-    }
-
-    @Test
     void updatePaymentAndSubmissionStatus_updatesPaymentAndSubmissionStatus_returnNothing() throws DissolutionNotFoundException {
         PaymentPatchRequest data = generatePaymentPatchRequest();
 
         service.handlePayment(data, APPLICATION_REFERENCE);
 
         verify(patcher).handlePayment(data, APPLICATION_REFERENCE);
-    }
-
-    @Test
-    void checkPatchDirectorConstraints_returnsErrorString_whenDirectorIsNotPendingApproval() throws DissolutionNotFoundException {
-        Dissolution dissolution = generateDissolution();
-
-        when(repository.findByCompanyNumber(COMPANY_NUMBER)).thenReturn(Optional.of(dissolution));
-        when(getter.isDirectorPendingApprovalForDissolution(OFFICER_ID, dissolution)).thenReturn(false);
-
-        Optional<String> error = service.checkPatchDirectorConstraints(COMPANY_NUMBER, OFFICER_ID, EMAIL);
-
-        assertEquals("Director is not pending approval", error.get());
-    }
-
-    @Test
-    void checkPatchDirectorConstraints_returnsErrorString_whenDirectorIsEmailIsNotApplicant() throws DissolutionNotFoundException {
-        Dissolution dissolution = generateDissolution();
-
-        when(repository.findByCompanyNumber(COMPANY_NUMBER)).thenReturn(Optional.of(dissolution));
-        when(getter.isDirectorPendingApprovalForDissolution(OFFICER_ID, dissolution)).thenReturn(true);
-        when(getter.doesEmailBelongToApplicant(EMAIL, dissolution)).thenReturn(false);
-
-        Optional<String> error = service.checkPatchDirectorConstraints(COMPANY_NUMBER, OFFICER_ID, EMAIL);
-
-        assertEquals("Only the applicant can update signatory", error.get());
-    }
-
-    @Test
-    void checkPatchDirectorConstraints_OptionalEmpty_ChecksPass() throws DissolutionNotFoundException {
-        Dissolution dissolution = generateDissolution();
-
-        when(repository.findByCompanyNumber(COMPANY_NUMBER)).thenReturn(Optional.of(dissolution));
-        when(getter.isDirectorPendingApprovalForDissolution(OFFICER_ID, dissolution)).thenReturn(true);
-        when(getter.doesEmailBelongToApplicant(EMAIL, dissolution)).thenReturn(true);
-
-        Optional<String> error = service.checkPatchDirectorConstraints(COMPANY_NUMBER, OFFICER_ID, EMAIL);
-
-        assertTrue(error.isEmpty());
-    }
-
-    @Test
-    void checkPatchDirectorConstraints_throwsException_WhenDissolutionNotFound() throws DissolutionNotFoundException {
-        when(repository.findByCompanyNumber(COMPANY_NUMBER)).thenReturn(Optional.empty());
-
-        DissolutionNotFoundException thrown = assertThrows(DissolutionNotFoundException.class, () -> {
-            service.checkPatchDirectorConstraints(COMPANY_NUMBER, OFFICER_ID, EMAIL);
-        });
-
-        assertNotNull(thrown);
     }
 }
